botdosconcursos/requirements.txt

python-telegram-bot==20.3
Flask==3.0.0
requests==2.31.0
